- About us view has been created. (#12)
- Added app image to using in splash and about us view.
- Updated app icon
