+++ app/src/main/java/com/poli/botanicalassistant/ui/splash/SplashActivity.kt
@@ -1,0 +1,37 @@
+package com.poli.botanicalassistant.ui.splash
+
+import android.content.Intent
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import androidx.appcompat.app.AppCompatActivity
+import com.poli.botanicalassistant.MainActivity
+import com.poli.botanicalassistant.databinding.ActivitySplashBinding
+
+class SplashActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val binding = ActivitySplashBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        handleSplashScreen()
+    }
+
+    private fun handleSplashScreen() {
+        Handler(Looper.getMainLooper()).postDelayed({
+            navigateToMainActivity()
+        }, SPLASH_TIME_OUT)
+    }
+
+    private fun navigateToMainActivity() {
+        val intent = Intent(this, MainActivity::class.java)
+        startActivity(intent)
+        finish()
+    }
+
+    companion object {
+        private const val SPLASH_TIME_OUT: Long = 1500
+    }
+}
