+++ app/src/main/res/layout/activity_splash.xml
@@ -2,22 +2,14 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/white"> <!-- Cambia el color de fondo según lo desees -->
+    android:fitsSystemWindows="true">
 
     <ImageView
         android:id="@+id/logo"
         android:layout_width="@dimen/splash_size"
         android:layout_height="@dimen/splash_size"
         android:layout_centerInParent="true"
-        android:src="@drawable/your_logo" /> <!-- Reemplaza 'your_logo' con el nombre de tu archivo de imagen -->
+        android:contentDescription="@string/fullscreen_icon_description"
+        android:src="@drawable/botanical_assistant_img" />
 
-    <TextView
-        android:id="@+id/countdown_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/logo"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="16dp"
-        android:textSize="24sp"
-        android:textColor="@android:color/black" />
 </RelativeLayout>
