+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -5,10 +5,6 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_home"
-            android:title="@string/menu_home" />
-        <item
             android:id="@+id/nav_garden"
             android:icon="@drawable/ic_menu_garden"
             android:title="@string/menu_garden" />
@@ -20,5 +16,9 @@
             android:id="@+id/nav_web"
             android:icon="@drawable/ic_menu_web"
             android:title="@string/menu_web" />
+        <item
+            android:id="@+id/nav_about_us"
+            android:icon="@drawable/ic_menu_about_us"
+            android:title="@string/menu_about_us" />
     </group>
 </menu>
