+++ app/src/main/res/values/dimens.xml
@@ -19,5 +19,5 @@
     <dimen name="item_elevation">4dp</dimen>
     <dimen name="splash_size">150dp</dimen>
     <dimen name="logo_width">95dp</dimen>
-    <dimen name="logo_height">94dp</dimen>
+    <dimen name="logo_height">95dp</dimen>
 </resources>
