+++ app/src/main/res/values/strings.xml
@@ -8,12 +8,13 @@
     <string name="action_settings">Settings</string>
 
     <string name="menu_home">Home</string>
+    <string name="menu_about_us">About us</string>
     <string name="menu_garden">Garden</string>
     <string name="menu_slideshow">Slideshow</string>
     <string name="menu_tutorials">Tutorials</string>
     <string name="menu_web">Web</string>
-    <string name="menu_video_player">Video Player</string> <!-- Añadido para el VideoPlayerFragment -->
-    <string name="menu_videos">Videos</string> <!-- Asegúrate de que esta línea esté presente -->
+    <string name="menu_video_player">Video Player</string>
+    <string name="menu_videos">Videos</string>
 
     <string name="blog_image_preview">Blog image content description</string>
     <string name="lab_input_url">URL</string>
@@ -25,8 +26,8 @@
     <string name="plant_classification_succulent">Succulent</string>
     <string name="plant_classification_other">Other</string>
 
-    <string name="video_image">Video image description</string> <!-- Descripción de la imagen del video -->
-    <string name="video_icon">Video icon</string> <!-- Descripción accesible para el ícono de video -->
+    <string name="video_image">Video image description</string>
+    <string name="video_icon">Video icon</string>
 
     <string name="video_player_error_generic_video">There seems to be an issue playing the video. Please try again.</string>
     <string name="video_player_error_video_not_found">The video you are trying to watch is not available. Please check the video ID.</string>
@@ -34,5 +35,22 @@
     <string name="video_player_error_unexpected">An unexpected error occurred while playing the video. Please try again later.</string>
 
     <string name="fullscreen_icon_description">Fullscreen</string>
+
+    <string name="about_app_description">Botanical Assistant helps you identify and take care of your plants with personalized recommendations and an extensive plant database.</string>
+    <string name="features_title">Main Features</string>
+    <string name="about_team">Developed by the Botanical Assistant Team</string>
+    <string name="privacy_policy">Privacy Policy</string>
+    <string name="terms_of_service">Terms of Service</string>
+    <string name="what_we_offer_title">What We Offer</string>
+    <string name="welcome_message">Welcome to our app, designed to provide everything you need to know about plants! Here’s what you can explore:</string>
+    <string name="plant_database_feature">Extensive Plant Database: Browse through a wide variety of plants, each accompanied by detailed information such as life cycles, care instructions, and categories.</string>
+    <string name="plant_care_guidance_feature">Plant Care Guidance: Discover how to properly nurture and care for your plants with personalized tips on watering, sunlight, and ideal temperatures.</string>
+    <string name="video_tutorials_feature">Video Tutorials: Access a curated collection of videos offering tutorials, recommendations, and expert advice on planting, care, and more.</string>
+    <string name="web_navigation_feature">Web Navigation: Easily search and explore web pages dedicated to plant-related content, helping you find exactly what you are looking for.</string>
+
+    <string name="mission_statement">
+    With these features, our mission is to empower you to take better care of your plants and help you grow your gardening knowledge.
+</string>
+
 
 </resources>
